+++ src/github.ts
@@ -128,21 +128,3 @@ export async function get_file_from_uri(email: string, uri: URL, req: Request, r
         return res.status(500).send('Internal Server Error');
     }
 }
-
-// stages of building project data are:
-// 0: basic project structure
-// 1: full project structure
-// 2: first 5 files + package.json (if exist)
-// 3: first 5 files + package.json (if exist, and using boostignore and gitignore)
-// 4: all file data (trimmed to ignore files)
-export async function user_project_data_references(uri: URL, stage: number, req: Request, res: Response) : Promise<string> {
-
-    if (stage < 0 || stage > 4) {
-        res.status(400).send('Invalid stage');
-        return "";
-    }
-
-    console.log(`Writing sample project data: ${uri}`);
-
-    return uri.toString();
-}